requirements.txt
advent-of-code-data
parse
matplotlib
dijkstar
pillow
